requirements.txt
# ===================================
# Air Quality RAG Chatbot
# Simple & Clean - No conflicts!
# ===================================

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Google Gemini AI
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Validation
pydantic==2.5.0

# CORS Support
python-multipart==0.0.6

# Vector Store and Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
huggingface-hub==0.16.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2